wlm: place base classes to s3 tests
