--- ydb/tests/workload_manager/s3/ya.make
@@ -1,12 +1,0 @@
-PY3_LIBRARY()
-
-PY_SRCS(
-    base.py
-)
-
-PEERDIR(
-    ydb/tests/functional/tpc/lib
-    ydb/tests/workload_manager/common
-)
-
-END()
